Add most of a CGRU implementation. Not yet tested.
I've been testing individual pieces, so I'm pretty sure what I've got so
far will mostly work. I suspect it will give awful results without
hyperparameter optimization, though, because they had to try hundreds of
models and only got a few that worked in the paper (!?). this might be
because they were trying it on multiplication, though, which is a specific
result that is either right or not, in contrast to next-word modeling
which is extremely fuzzy.